pyproject.toml
[tool.poetry]
name = "branda-server"
version = "0.1.0"
description = ""
authors = ["litmus-zhang <abdulsalamlukmon9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.3"
langchain-community = "^0.2.4"
langchain-core = "^0.2.5"
langchain-google-vertexai = "^1.0.5"
langchain-google-genai = "^1.0.6"
markdown = "^3.6"
langchain-openai = "^0.1.8"
python-dotenv = "^1.0.1"
fastapi = {extras = ["standard"], version = "^0.112.2"}
uvicorn = {extras = ["standard"], version = "^0.30.6"}


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.10.0"
pylint = "^3.2.3"
mimesis = "^17.0.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
faker = "^25.8.0"
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
